+++ PostHog/Surveys/PostHogSurveyIntegration.swift
@@ -42,6 +42,7 @@
 
         private var didBecomeActiveToken: RegistrationToken?
         private var didLayoutViewToken: RegistrationToken?
+        private var eventCapturedToken: RegistrationToken?
 
         private var activeSurveyLock = NSLock()
         private var activeSurvey: PostHogSurvey?
@@ -73,6 +74,11 @@
 
         func start() {
             #if os(iOS)
+                // Subscribe to event captures
+                eventCapturedToken = postHog?.onEventCaptured.subscribe { [weak self] event in
+                    self?.onEvent(event: event)
+                }
+
                 // TODO: listen to screen view events
                 didLayoutViewToken = DI.main.viewLayoutPublisher.onViewLayout.subscribe(throttle: 5) { [weak self] in
                     self?.showNextSurvey()
@@ -85,6 +91,7 @@
         }
 
         func stop() {
+            eventCapturedToken = nil
             didBecomeActiveToken = nil
             didLayoutViewToken = nil
             #if os(iOS)
@@ -111,8 +118,7 @@
                     .filter { survey in // 3. and match display conditions,
                         // TODO: Check screen conditions
                         // TODO: Check event conditions
-                        let deviceTypeCheck = self.doesSurveyDeviceTypesMatch(survey: survey)
-                        return deviceTypeCheck
+                        self.doesSurveyDeviceTypesMatch(survey: survey)
                     }
                     .filter { survey in // 3.5. wait period has passed
                         self.hasWaitPeriodPassed(survey: survey)
@@ -143,14 +149,12 @@
             }
         }
 
-        // TODO: Decouple PostHogSDK and use registration handlers instead
-        /// Called from PostHogSDK instance when an event is captured
-        func onEvent(event: String, properties: [String: Any]) {
-            let candidates = eventsToSurveysLock.withLock { eventsToSurveys[event] } ?? []
+        private func onEvent(event: PostHogEvent) {
+            let candidates = eventsToSurveysLock.withLock { eventsToSurveys[event.event] } ?? []
             guard !candidates.isEmpty else { return }
 
             let matchingSurveyIds = candidates
-                .filter { matchPropertyFilters($0.condition.propertyFilters, eventProperties: properties) }
+                .filter { matchPropertyFilters($0.condition.propertyFilters, eventProperties: event.properties) }
                 .map(\.surveyId)
 
             guard !matchingSurveyIds.isEmpty else { return }
@@ -331,9 +335,7 @@
 
             let key = getSurveySeenKey(survey)
             let surveysSeen = getSeenSurveyKeys()
-            let surveySeen = surveysSeen[key] as? Bool ?? false
-
-            return surveySeen
+            return surveysSeen[key] as? Bool ?? false
         }
 
         /// Mark a survey as seen
@@ -1025,6 +1027,10 @@
                 eventActivatedSurveysLock.withLock {
                     eventActivatedSurveys.contains(surveyId)
                 }
+            }
+
+            func testOnEvent(event: PostHogEvent) {
+                onEvent(event: event)
             }
 
             static func clearInstalls() {